@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,6 +33,7 @@ dependencies:
   firebase_core_web: ^1.7.1
   cloud_firestore: ^3.4.3
   url_launcher: ^6.0.4
+  firebase_database: ^9.1.0
   http: ^0.13.3
   flutter:
     sdk: flutter
